Update colors, add fonts, remove store
This commit is the first in a series of commits that aim to simplify the
site, and refresh the look and feel.

Step one is removing the e-commerce functionality from the site.
WooCommerce is purged from the site in this commit.

Some styling and design changes are also present, but nothing is
finalized yet.