Add serverless function for form submissions
Address a security vulnerability by making requests to an external API
in a serverless function.

The proof-of-concept for this feature leaked the API token, and the
token would be accessible by the client. To address this issue, the API
token is only accessed by the serverless function, and is no longer
passed to the client.