Pin dependencies
Update and pin all dependencies to their current (latest) versions.

Going forward, avoid using version constraints that permit a range of
package versions to be installed. This project has very few direct
dependencies, and the best way to keep everything up-to-date and secure
is always run on the latest version of every installed package. The
assumption is that the latest versions will include the most up-to-date
security fixes, and by pinning each package version, the package manager
won't be able to install any versions of a package that are outdated,
which could contain security vulnerabilities.

In practice, this is a theoretical ideal that doesn't always work out.
One drawback of this is that pinning each package to a single version
may cause compatibility problems among dependencies. In these cases, it
will be necessary to add more permissive version constraints, which
could then be removed in the future when the compatibility issues have
been resolved.