Ensure `hugo_stats.json` is written in dev
Update Hugo config files to ensure `hugo_stats.json` is written in
development.

This project uses the PurgeCSS plugin to eliminate unused CSS from the
asset bundle. The purging of unused CSS occurs in development and
production environments for an important reason: if PurgeCSS is not
running in development, it is extremely easy to deploy code to
production with missing CSS, because the CSS that is served in
development is un-purged (and therefore completely different) than the
CSS that is served in production.

The `hugo_stats.json` file contains a list of HTML elements that
PurgeCSS uses to determine what CSS needs to be removed, and what CSS
rules are being used. This file must be updated every time the site is
rebuilt in development in order for PurgeCSS to work correctly.