Remove XXL-breakpoint-specific column offset
Change page layouts by removing an two-column offset on XXL viewports.
This change does create more white space, which is particularly
noticeable on pages with little text content. However, I think this is
better—with the XXL offset, context would be squeezed together even more
than on when viewed on an XL viewport.